% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_condition.R
\name{capture_message}
\alias{capture_message}
\alias{capture_warning}
\alias{capture_condition}
\title{capture condition}
\usage{
capture_message(expr)

capture_warning(expr)
}
\description{
capture condition
}
\examples{
foom <- function(x) {
  message("its too bad")
  return(x)
}
capture_message(foom(4))

foow <- function(x) {
  warning("its too bad")
  return(x)
}
capture_warning(foow(4))
}
\keyword{internal}
