% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionz-package.R
\docType{package}
\name{conditionz-package}
\alias{conditionz-package}
\alias{conditionz}
\title{condition control}
\author{
Scott Chamberlain \href{mailto:myrmecocystus@gmail.com}{myrmecocystus@gmail.com}
}
\keyword{package}
