% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionKeeper.R
\docType{data}
\name{ConditionKeeper}
\alias{ConditionKeeper}
\title{ConditionKeeper}
\arguments{
\item{times}{(integer) number of times to throw condition. required.
default: 1}
}
\description{
ConditionKeeper
}
\details{
\strong{Methods}
\itemize{
\item \code{add()} - xx
\item \code{remove()} - xx
\item \code{purge()} - xx
\item \code{thrown_already()} - xx
\item \code{not_thrown_yet()} - xx
\item \code{thrown_times()} - xx
\item \code{thrown_enough()} - xx
\item \code{get_id()} - xx
\item \code{handle_conditions()} - xx
}
}
\examples{
x <- ConditionKeeper$new(times = 4)
x
x$get_id()
x$add("one")
x$add("two")
x
x$thrown_already("one")
x$thrown_already("bears")
x$not_thrown_yet("bears")

x$add("two")
x$add("two")
x$add("two")
x$thrown_times("two")
x$thrown_enough("two")
x$thrown_enough("one")

foo <- function(x) {
  message("you gave: ", x)
  return(x)
}
foo('a')
x$handle_conditions(foo('a'))

x <- ConditionKeeper$new(times = 4, condition = "warning")
x
x$add("one")
x$add("two")
x
}
\keyword{datasets}
